storing records in recordhook
using eval to find globals is typically bad practice and difficult to maintain. This adds a RecordHook metaclass that gets called anytime Record is subclassed. The name of the class is saved in RecordHook.records as a dict where it can be accessed from anywhere without using eval or global