refactor field_value_array
`field_value_array` was using`field_value` to find the array of field values, which is probably the wrong thing to do. This finds all field_values first, then filters based on the name and role. If the field_type of the operation's OperationType is not an array, it raises an error.